backend/requirements.txt
django<6.0
djangorestframework
mysqlclient
python-dotenv
django-cors-headers
djangorestframework-simplejwt
drf-spectacular
